native with local db and server/server/backend/requirements.txt
Django==3.1
djangorestframework==3.11.1
django-cors-headers==3.4.0
pillow==7.2.0
